coin-classification/requirements.txt
torch
transformers
numpy
pandas
scikit-learn
tqdm
matplotlib
seaborn
Pillow
